+++ app/controllers/hands_controller.rb
@@ -1,5 +1,6 @@
 class HandsController < ApplicationController
   protect_from_forgery except: :index
+  include PokerJudgeService
 
   def index
     if request.post?
